Move method away from _init_
